# [Core][GeoMechanicsApplication] Add a test suite without a kernel to improve performance of GTest runs.