--- a/applications/GeoMechanicsApplication/tests/cpp_tests/geo_mechanics_fast_suite.h
+++ b/applications/GeoMechanicsApplication/tests/cpp_tests/geo_mechanics_fast_suite.h
@@ -30,6 +30,10 @@ namespace Kratos::Testing
      KratosLinearSolversApplication::Pointer mpLinearSolversApp;
  };
 
+class KratosGeoMechanicsFastSuiteWithoutKernel : public KratosCoreFastSuiteWithoutKernel
+{
+};
+
  class KratosGeoMechanicsIntegrationSuite : public KratosCoreFastSuite
  {
  public: