**📝 Description**
One of the characteristics of running our unit tests using the GTest suites is that currently every test has a kernel and initializes th applications. However, for the GeoMechanicsApplication, most unit tests are independent of the initialization of the apps, since they test the units in isolation. Therefore, we can achieve a significant speed-up by having a version of the suites without a kernel. We experimented with this idea and could reduce the unit test time for the Geo app by > 75% (~35 sec to ~7.5 sec locally) using this mechanism.

This PR adds a kernel-less test suite to Kratos Core, which is then used as an example in the GeoMechanicsApplication. In this PR one example file is converted to the suite without kernel (for a single test locally, the speed-up is from ~100 ms to < 1 ms)